+++ addon/components/paper-input.js
@@ -1,12 +1,14 @@
 import Ember from 'ember';
 import BaseFocusable from './base-focusable';
 import ColorMixin from 'ember-paper/mixins/color-mixin';
+import FlexMixin from 'ember-paper/mixins/flex-mixin';
 
-export default BaseFocusable.extend(ColorMixin, {
+export default BaseFocusable.extend(ColorMixin, FlexMixin, {
   tagName: 'md-input-container',
   classNames: ['md-default-theme'],
   classNameBindings: ['hasValue:md-input-has-value', 'focus:md-input-focused', 'isInvalid:md-input-invalid'],
   type: 'text',
+  autofocus: false,
   tabindex: -1,
   hasValue: Ember.computed.notEmpty('value'),
   inputElementId: Ember.computed('elementId', function() {
@@ -63,11 +65,18 @@ export default BaseFocusable.extend(ColorMixin, {
   },
 
   actions: {
-    focusIn() {
+    focusIn(value) {
+      // We resend action so other components can take use of the actions also ( if they want ).
+      // Actions must be sent before focusing.
+      this.sendAction('focus-in', value);
       this.set('focus',true);
     },
-    focusOut() {
+    focusOut(value) {
+      this.sendAction('focus-out', value);
       this.set('focus',false);
+    },
+    keyDown(value, event) {
+      this.sendAction('key-down', value, event);
     }
   }
 });